server/requirements.txt
CherryPy==3.6.0
gunicorn==19.5.0
celery==4.2.1
kombu==4.4.0
aiohttp==0.12.0
psutil==5.6.6
pyowm==2.10.0 
feedparser==5.2.1 
jsonpickle==1.1
python-dateutil==2.8.0 
PyMySQL==0.9.3
spotipy==2.4.4
wikipedia==1.4.0
